Domain, Test Problem, and First Problem
The Domain includes inspect, patch, and weld actions. These actions must be carried out by egobots. Inspect can also be carried out by sidekicks. The Domain also includes locations and adjacencies which allow different paths of movement for egobots and sidekicks. The idea is that sidekicks can move everywhere, so egobots need sidekicks to bring them tools for patching and welding. Two tools exist: patches, which are consumed when used to patch a panel; and welders, which are not consumed when used to weld a panel.