Added Hands-Free to goals as a social norm
If each egobot doesn't have a goal of hands-free, they're going to hold onto welders when they're done with them. There is no intended functionality for the sidekick to request actions from the egobots, so this would result in a failure. Hands-free being in the goal state acts like a social norm that the egobots all obey.