Renaming
Egobots can use the same system. Renamed global effects to events, these are events the agent listens for when parsing, regardless of which agent carries them out. Actions and Events could be bundled together if we have an arbitrarily long list of identifiers and I can parse for all of them at once.